requirements.txt
# Core dependencies for YouTube video processing
yt-dlp>=2025.8.11
youtube-transcript-api>=1.2.2
python-dotenv>=1.0.0

# LangChain for AI model integration
langchain>=0.3.27
langchain-openai>=0.3.30
langchain-anthropic>=0.3.18
langchain-community>=0.3.27

# Export utilities
reportlab>=4.4.3  # For PDF export
markdown>=3.4.4   # For HTML export
requests>=2.31.0
beautifulsoup4>=4.12.0  # For fallback metadata extraction

# Optional: For local model support with Ollama
# Install ollama separately if needed

# Telegram Bot Integration
python-telegram-bot>=20.0

# Template engine for V2 redesign
jinja2>=3.1
markupsafe>=2.1

# HTML sanitization for V2 security
bleach>=6.1

# PostgreSQL support for migration
psycopg2-binary>=2.9.0

# Google OIDC verification for multi-user auth
google-auth>=2.16.0

# Clerk JWT verification for Jeop3 game creation auth
pyjwt>=2.9.0
